requirements.txt
alabaster==0.7.13
Babel==2.13.1
certifi==2023.7.22
cffi==1.16.0
charset-normalizer==3.3.1
docutils==0.20.1
idna==3.4
imagesize==1.4.1
Jinja2==3.1.2
MarkupSafe==2.1.3
numpy==1.26.1
opencv-python==4.8.1.78
packaging==23.2
pdoc==14.1.0
PyAudio==0.2.13
pycparser==2.21
pyee==11.0.1
Pygments==2.16.1
python-ffmpeg==2.0.4
requests==2.31.0
snowballstemmer==2.2.0
sounddevice==0.4.6
sphinxcontrib-applehelp==1.0.7
sphinxcontrib-devhelp==1.0.5
sphinxcontrib-htmlhelp==2.0.4
sphinxcontrib-jsmath==1.0.1
sphinxcontrib-qthelp==1.0.6
sphinxcontrib-serializinghtml==1.1.9
typing_extensions==4.8.0
urllib3==2.0.7
